Split the BootStrap.groovy file to two files
Code that should run in applications where the plugin is installed is
now in a custom Bootstrap class so it will get picked up and run.

The test data on the other hand remains in the default BootStrap file as
that will not get packaged into the proper plugin release.