Added support for live reloading in development mode
Annotations can now be added to or removed from domain class properties
and the plugin shall react to the changes accordingly. As there seems to
be no way to unregister an ObjectMarshaller an obsoleted instance will
now simply throw an exception instead of continuing working until the
next app restart.